Create PowerCards Replacements for Melee attacks
This file has the 'Description & Notes' field of a handout. This handout's name MUST begin with "PowerCards Replacements" to work with PowerCards.
The replacements here cover all the SRD melee weapons and work with the Spear.txt macro. To use that macro, make sure this handout exists and simply change the line in the macro where it says "Spear" to whatever weapon you want.